+++ digdag-standards/src/main/java/io/digdag/standards/operator/param/ParamGetOperatorFactory.java
@@ -21,7 +21,7 @@ public class ParamGetOperatorFactory
     private DBI dbi;
 
     @Inject
-    public ParamGetOperatorFactory(Config systemConfig, @Named("user_database") DBI dbi)
+    public ParamGetOperatorFactory(Config systemConfig, @Named("param_server.database") DBI dbi)
     {
         this.systemConfig = systemConfig;
         this.dbi = dbi;
