+++ digdag-standards/src/main/java/io/digdag/standards/operator/param/ParamServerModule.java
@@ -5,9 +5,6 @@ import com.google.inject.Module;
 import com.google.inject.Scopes;
 import com.google.inject.name.Names;
 import io.digdag.core.database.DatabaseConfig;
-import io.digdag.core.database.DatabaseModule;
-import io.digdag.core.database.UserDataSourceProvider;
-import io.digdag.core.database.UserDatabaseConfigProvider;
 import org.skife.jdbi.v2.DBI;
 
 import javax.sql.DataSource;
@@ -20,13 +17,13 @@ public class ParamServerModule
     public void configure(Binder binder)
     {
         binder.bind(DatabaseConfig.class)
-                .annotatedWith(Names.named("user_database"))
+                .annotatedWith(Names.named("param_server.database"))
                 .toProvider(UserDatabaseConfigProvider.class).in(Scopes.SINGLETON);
         binder.bind(DataSource.class)
-                .annotatedWith(Names.named("user_database"))
+                .annotatedWith(Names.named("param_server.database"))
                 .toProvider(UserDataSourceProvider.class).in(Scopes.SINGLETON);
         binder.bind(DBI.class)
-                .annotatedWith(Names.named("user_database"))
-                .toProvider(DatabaseModule.UserDbiProvider.class);  // don't make this singleton because DBI.registerMapper is called for each StoreManager
+                .annotatedWith(Names.named("param_server.database"))
+                .toProvider(UserDbiProvider.class);  // don't make this singleton because DBI.registerMapper is called for each StoreManager
     }
 }
